Bayesian Sampler Model

As described in (Zhu et al. 2020) . Vectors can be provided for each parameter, allowing multiple estimates at once.

Bayesian_Sampler(
  a_and_b,
  b_and_not_a,
  a_and_not_b,
  not_a_and_not_b,
  beta,
  N,
  N2 = NULL
)

Arguments

a_and_b, b_and_not_a, a_and_not_b, not_a_and_not_b

True probabilites for the conjuctions and disjunctions of A and B. Must add to 1.

beta

Prior parameter.

N

Number of samples drawn

N2

Optional. Number of samples drawn for conjunctions and disjunctions. (called N' in the paper). If not given, it will default to N2=N. Must be equal or smaller than N.

Value

Named list with predicted probabilities for every possible combination of A and B.

Examples

Bayesian_Sampler(
    a_and_b = c(.4, .25),
    b_and_not_a = c(.4,  .25),
    a_and_not_b = c(.1, .25),
    not_a_and_not_b = c(.1, .25),
    beta = 1,
    N <- c(10, 12),
    N2 <- c(10, 10)
)
#> $a_and_b
#> [1] 0.4166667 0.2916667
#> 
#> $b_and_not_a
#> [1] 0.4166667 0.2916667
#> 
#> $a_and_not_b
#> [1] 0.1666667 0.2916667
#> 
#> $not_a_and_not_b
#> [1] 0.1666667 0.2916667
#> 
#> $a
#> [1] 0.5 0.5
#> 
#> $b
#> [1] 0.75 0.50
#> 
#> $not_a
#> [1] 0.5 0.5
#> 
#> $not_b
#> [1] 0.25 0.50
#> 
#> $a_or_b
#> [1] 0.8333333 0.7083333
#> 
#> $a_or_not_b
#> [1] 0.8333333 0.7083333
#> 
#> $b_or_not_a
#> [1] 0.5833333 0.7083333
#> 
#> $not_a_or_not_b
#> [1] 0.5833333 0.7083333
#> 
#> $a_given_b
#> [1] 0.5 0.5
#> 
#> $not_a_given_b
#> [1] 0.5 0.5
#> 
#> $a_given_not_b
#> [1] 0.5 0.5
#> 
#> $not_a_given_not_b
#> [1] 0.5 0.5
#> 
#> $b_given_a
#> [1] 0.75 0.50
#> 
#> $not_b_given_a
#> [1] 0.25 0.50
#> 
#> $b_given_not_a
#> [1] 0.75 0.50
#> 
#> $not_b_given_not_a
#> [1] 0.25 0.50
#>