Metropolis-coupled MCMC sampler (MC3)

This sampler is a variant of MCMC in which multiple parallel chains are run at different temperatures. The chains stochastically swap positions which allows the coldest chain to visit regions far from its starting point (unlike in MCMC). Because of this, an MC3 sampler can explore far-off regions, whereas an MCMC sampler may become stuck in a particular point of high density.

sampler_mc3(
  distr_name,
  distr_params,
  start,
  sigma_prop = NULL,
  nChains = 6,
  delta_T = 4,
  swap_all = TRUE,
  iterations = 1024L,
  weights = NULL
)

Arguments

distr_name
Name of the distribution from which to sample from.

distr_params
Distribution parameters.

start
Vector. Starting position of the sampler.

sigma_prop
Covariance matrix of the proposal distribution. If sampling in 1D space, it can be instead a number.

nChains
Number of chains to run.

delta_T
numeric, >1. Temperature increment parameter. The bigger this number, the steeper the increase in temperature between the cold chain and the next chain

swap_all
Boolean. If true, every iteration attempts floor(nChains / 2) swaps. If false, only one swap per iteration.

iterations
Number of iterations of the sampler.

weights
If using a mixture distribution, the weights given to each constituent distribution. If none given, it defaults to equal weights for all distributions.

Examples

# Sample from a normal distribution
mc_3 <- sampler_mc3(distr_name = "norm", distr_params = c(0,1), start = 1, sigma_prop = diag(1))