Bayesian Sampler Model

As described in (Zhu et al. 2020) . Vectors can be provided for each parameter, allowing multiple estimates at once.

Bayesian_Sampler(
  a_and_b,
  b_and_not_a,
  a_and_not_b,
  not_a_and_not_b,
  beta,
  N,
  N2 = NULL,
  return = "mean",
  n_simulations = 1000
)

Arguments

a_and_b, b_and_not_a, a_and_not_b, not_a_and_not_b

True probabilites for the conjuctions and disjunctions of A and B. Must add to 1.

beta

Prior parameter.

N

Number of samples drawn

N2

Optional. Number of samples drawn for conjunctions and disjunctions. (called N' in the paper). If not given, it will default to N2=N. Must be equal or smaller than N.

return

Optional. Either "mean", "variance" or "simulation". Defaults to "mean".

n_simulations

Optional. if return="simulation", how many simulations per possible combination of A and B. Defaults to 1000.

Value

If return="mean" or return="variance", named list with predicted probabilities for every possible combination of A and B, or the expected variance of those predictions. If return="simulation", simulated predictions instead. Note that if return="simulation", the named list will contain vectors if the length of the true probabilities is 1; otherwise a matrix where each column is a queried probability and each row a simulation

References

Zhu J, Sanborn AN, Chater N (2020). “The Bayesian Sampler: Generic Bayesian Inference Causes Incoherence in Human Probability Judgments.” Psychological Review, 127(5), 719--748. doi:10.1037/rev0000190 .

Examples

Bayesian_Sampler(
    a_and_b = c(.4, .25),
    b_and_not_a = c(.4,  .25),
    a_and_not_b = c(.1, .25),
    not_a_and_not_b = c(.1, .25),
    beta = 1,
    N <- c(10, 12),
    N2 <- c(10, 10)
)
#> $a_and_b
#> [1] 0.4166667 0.2916667
#> 
#> $b_and_not_a
#> [1] 0.4166667 0.2916667
#> 
#> $a_and_not_b
#> [1] 0.1666667 0.2916667
#> 
#> $not_a_and_not_b
#> [1] 0.1666667 0.2916667
#> 
#> $a
#> [1] 0.5 0.5
#> 
#> $b
#> [1] 0.75 0.50
#> 
#> $not_a
#> [1] 0.5 0.5
#> 
#> $not_b
#> [1] 0.25 0.50
#> 
#> $a_or_b
#> [1] 0.8333333 0.7083333
#> 
#> $a_or_not_b
#> [1] 0.8333333 0.7083333
#> 
#> $b_or_not_a
#> [1] 0.5833333 0.7083333
#> 
#> $not_a_or_not_b
#> [1] 0.5833333 0.7083333
#> 
#> $a_given_b
#> [1] 0.5 0.5
#> 
#> $not_a_given_b
#> [1] 0.5 0.5
#> 
#> $a_given_not_b
#> [1] 0.5 0.5
#> 
#> $not_a_given_not_b
#> [1] 0.5 0.5
#> 
#> $b_given_a
#> [1] 0.75 0.50
#> 
#> $not_b_given_a
#> [1] 0.25 0.50
#> 
#> $b_given_not_a
#> [1] 0.75 0.50
#> 
#> $not_b_given_not_a
#> [1] 0.25 0.50
#> 
Bayesian_Sampler(
   a_and_b = c(0.05, .85),
   b_and_not_a = c(.85,  0.05),
   a_and_not_b = c(.05, 0.05),
   not_a_and_not_b = c(0.05, 0.05),
   beta = 1,
   N = 5,
   return="simulation"
)$a
#>              [,1]      [,2]
#>    [1,] 0.1428571 0.8571429
#>    [2,] 0.7142857 0.8571429
#>    [3,] 0.1428571 0.8571429
#>    [4,] 0.2857143 0.5714286
#>    [5,] 0.2857143 0.8571429
#>    [6,] 0.1428571 0.7142857
#>    [7,] 0.4285714 0.8571429
#>    [8,] 0.2857143 0.8571429
#>    [9,] 0.1428571 0.5714286
#>   [10,] 0.1428571 0.8571429
#>   [11,] 0.1428571 0.8571429
#>   [12,] 0.2857143 0.7142857
#>   [13,] 0.1428571 0.7142857
#>   [14,] 0.2857143 0.8571429
#>   [15,] 0.1428571 0.8571429
#>   [16,] 0.1428571 0.8571429
#>   [17,] 0.1428571 0.8571429
#>   [18,] 0.1428571 0.8571429
#>   [19,] 0.1428571 0.5714286
#>   [20,] 0.2857143 0.8571429
#>   [21,] 0.1428571 0.5714286
#>   [22,] 0.2857143 0.8571429
#>   [23,] 0.1428571 0.7142857
#>   [24,] 0.2857143 0.8571429
#>   [25,] 0.4285714 0.7142857
#>   [26,] 0.2857143 0.7142857
#>   [27,] 0.1428571 0.7142857
#>   [28,] 0.1428571 0.8571429
#>   [29,] 0.1428571 0.8571429
#>   [30,] 0.1428571 0.7142857
#>   [31,] 0.1428571 0.8571429
#>   [32,] 0.2857143 0.7142857
#>   [33,] 0.1428571 0.7142857
#>   [34,] 0.2857143 0.8571429
#>   [35,] 0.2857143 0.7142857
#>   [36,] 0.2857143 0.8571429
#>   [37,] 0.2857143 0.5714286
#>   [38,] 0.1428571 0.7142857
#>   [39,] 0.1428571 0.8571429
#>   [40,] 0.1428571 0.8571429
#>   [41,] 0.1428571 0.8571429
#>   [42,] 0.2857143 0.7142857
#>   [43,] 0.1428571 0.8571429
#>   [44,] 0.2857143 0.7142857
#>   [45,] 0.1428571 0.8571429
#>   [46,] 0.2857143 0.7142857
#>   [47,] 0.2857143 0.7142857
#>   [48,] 0.1428571 0.8571429
#>   [49,] 0.1428571 0.8571429
#>   [50,] 0.2857143 0.8571429
#>   [51,] 0.2857143 0.8571429
#>   [52,] 0.1428571 0.8571429
#>   [53,] 0.2857143 0.8571429
#>   [54,] 0.4285714 0.7142857
#>   [55,] 0.1428571 0.5714286
#>   [56,] 0.1428571 0.7142857
#>   [57,] 0.2857143 0.7142857
#>   [58,] 0.1428571 0.7142857
#>   [59,] 0.1428571 0.8571429
#>   [60,] 0.2857143 0.7142857
#>   [61,] 0.1428571 0.8571429
#>   [62,] 0.1428571 0.7142857
#>   [63,] 0.2857143 0.8571429
#>   [64,] 0.1428571 0.8571429
#>   [65,] 0.1428571 0.8571429
#>   [66,] 0.4285714 0.7142857
#>   [67,] 0.2857143 0.8571429
#>   [68,] 0.2857143 0.7142857
#>   [69,] 0.2857143 0.8571429
#>   [70,] 0.2857143 0.7142857
#>   [71,] 0.2857143 0.7142857
#>   [72,] 0.2857143 0.8571429
#>   [73,] 0.2857143 0.8571429
#>   [74,] 0.1428571 0.7142857
#>   [75,] 0.1428571 0.8571429
#>   [76,] 0.1428571 0.7142857
#>   [77,] 0.1428571 0.5714286
#>   [78,] 0.2857143 0.7142857
#>   [79,] 0.2857143 0.7142857
#>   [80,] 0.1428571 0.8571429
#>   [81,] 0.2857143 0.7142857
#>   [82,] 0.2857143 0.7142857
#>   [83,] 0.1428571 0.8571429
#>   [84,] 0.1428571 0.8571429
#>   [85,] 0.2857143 0.8571429
#>   [86,] 0.1428571 0.8571429
#>   [87,] 0.2857143 0.8571429
#>   [88,] 0.1428571 0.8571429
#>   [89,] 0.2857143 0.4285714
#>   [90,] 0.2857143 0.8571429
#>   [91,] 0.1428571 0.8571429
#>   [92,] 0.1428571 0.7142857
#>   [93,] 0.4285714 0.7142857
#>   [94,] 0.1428571 0.8571429
#>   [95,] 0.1428571 0.8571429
#>   [96,] 0.1428571 0.8571429
#>   [97,] 0.1428571 0.8571429
#>   [98,] 0.2857143 0.7142857
#>   [99,] 0.2857143 0.8571429
#>  [100,] 0.2857143 0.5714286
#>  [101,] 0.2857143 0.7142857
#>  [102,] 0.1428571 0.8571429
#>  [103,] 0.2857143 0.8571429
#>  [104,] 0.2857143 0.8571429
#>  [105,] 0.1428571 0.7142857
#>  [106,] 0.4285714 0.7142857
#>  [107,] 0.1428571 0.7142857
#>  [108,] 0.1428571 0.7142857
#>  [109,] 0.1428571 0.8571429
#>  [110,] 0.2857143 0.8571429
#>  [111,] 0.1428571 0.8571429
#>  [112,] 0.4285714 0.7142857
#>  [113,] 0.1428571 0.7142857
#>  [114,] 0.1428571 0.8571429
#>  [115,] 0.1428571 0.8571429
#>  [116,] 0.1428571 0.7142857
#>  [117,] 0.2857143 0.8571429
#>  [118,] 0.1428571 0.7142857
#>  [119,] 0.1428571 0.7142857
#>  [120,] 0.1428571 0.8571429
#>  [121,] 0.1428571 0.8571429
#>  [122,] 0.2857143 0.8571429
#>  [123,] 0.2857143 0.8571429
#>  [124,] 0.2857143 0.5714286
#>  [125,] 0.1428571 0.7142857
#>  [126,] 0.2857143 0.8571429
#>  [127,] 0.2857143 0.8571429
#>  [128,] 0.1428571 0.7142857
#>  [129,] 0.1428571 0.8571429
#>  [130,] 0.1428571 0.7142857
#>  [131,] 0.1428571 0.7142857
#>  [132,] 0.1428571 0.8571429
#>  [133,] 0.1428571 0.8571429
#>  [134,] 0.1428571 0.7142857
#>  [135,] 0.1428571 0.7142857
#>  [136,] 0.1428571 0.8571429
#>  [137,] 0.1428571 0.8571429
#>  [138,] 0.1428571 0.5714286
#>  [139,] 0.2857143 0.7142857
#>  [140,] 0.5714286 0.8571429
#>  [141,] 0.1428571 0.7142857
#>  [142,] 0.1428571 0.8571429
#>  [143,] 0.1428571 0.8571429
#>  [144,] 0.2857143 0.7142857
#>  [145,] 0.1428571 0.7142857
#>  [146,] 0.1428571 0.7142857
#>  [147,] 0.2857143 0.8571429
#>  [148,] 0.1428571 0.5714286
#>  [149,] 0.2857143 0.8571429
#>  [150,] 0.4285714 0.8571429
#>  [151,] 0.2857143 0.8571429
#>  [152,] 0.2857143 0.7142857
#>  [153,] 0.2857143 0.8571429
#>  [154,] 0.1428571 0.8571429
#>  [155,] 0.2857143 0.8571429
#>  [156,] 0.1428571 0.7142857
#>  [157,] 0.2857143 0.8571429
#>  [158,] 0.4285714 0.7142857
#>  [159,] 0.2857143 0.7142857
#>  [160,] 0.4285714 0.8571429
#>  [161,] 0.1428571 0.8571429
#>  [162,] 0.1428571 0.8571429
#>  [163,] 0.1428571 0.7142857
#>  [164,] 0.2857143 0.7142857
#>  [165,] 0.2857143 0.8571429
#>  [166,] 0.2857143 0.7142857
#>  [167,] 0.2857143 0.8571429
#>  [168,] 0.1428571 0.8571429
#>  [169,] 0.1428571 0.7142857
#>  [170,] 0.1428571 0.7142857
#>  [171,] 0.2857143 0.7142857
#>  [172,] 0.4285714 0.8571429
#>  [173,] 0.4285714 0.8571429
#>  [174,] 0.1428571 0.8571429
#>  [175,] 0.1428571 0.8571429
#>  [176,] 0.1428571 0.5714286
#>  [177,] 0.4285714 0.8571429
#>  [178,] 0.1428571 0.7142857
#>  [179,] 0.4285714 0.8571429
#>  [180,] 0.2857143 0.8571429
#>  [181,] 0.2857143 0.7142857
#>  [182,] 0.1428571 0.8571429
#>  [183,] 0.1428571 0.8571429
#>  [184,] 0.1428571 0.7142857
#>  [185,] 0.4285714 0.8571429
#>  [186,] 0.2857143 0.8571429
#>  [187,] 0.1428571 0.5714286
#>  [188,] 0.2857143 0.7142857
#>  [189,] 0.2857143 0.7142857
#>  [190,] 0.2857143 0.8571429
#>  [191,] 0.2857143 0.8571429
#>  [192,] 0.1428571 0.7142857
#>  [193,] 0.2857143 0.8571429
#>  [194,] 0.1428571 0.8571429
#>  [195,] 0.1428571 0.7142857
#>  [196,] 0.2857143 0.8571429
#>  [197,] 0.1428571 0.8571429
#>  [198,] 0.2857143 0.8571429
#>  [199,] 0.1428571 0.8571429
#>  [200,] 0.2857143 0.8571429
#>  [201,] 0.1428571 0.8571429
#>  [202,] 0.1428571 0.8571429
#>  [203,] 0.2857143 0.7142857
#>  [204,] 0.1428571 0.7142857
#>  [205,] 0.5714286 0.7142857
#>  [206,] 0.2857143 0.8571429
#>  [207,] 0.5714286 0.7142857
#>  [208,] 0.2857143 0.8571429
#>  [209,] 0.4285714 0.7142857
#>  [210,] 0.2857143 0.8571429
#>  [211,] 0.1428571 0.7142857
#>  [212,] 0.1428571 0.8571429
#>  [213,] 0.1428571 0.8571429
#>  [214,] 0.1428571 0.7142857
#>  [215,] 0.2857143 0.5714286
#>  [216,] 0.1428571 0.8571429
#>  [217,] 0.1428571 0.7142857
#>  [218,] 0.1428571 0.5714286
#>  [219,] 0.2857143 0.7142857
#>  [220,] 0.2857143 0.8571429
#>  [221,] 0.2857143 0.8571429
#>  [222,] 0.1428571 0.8571429
#>  [223,] 0.1428571 0.8571429
#>  [224,] 0.2857143 0.8571429
#>  [225,] 0.1428571 0.7142857
#>  [226,] 0.4285714 0.8571429
#>  [227,] 0.2857143 0.8571429
#>  [228,] 0.2857143 0.7142857
#>  [229,] 0.2857143 0.5714286
#>  [230,] 0.1428571 0.8571429
#>  [231,] 0.1428571 0.8571429
#>  [232,] 0.1428571 0.8571429
#>  [233,] 0.1428571 0.8571429
#>  [234,] 0.2857143 0.7142857
#>  [235,] 0.2857143 0.7142857
#>  [236,] 0.2857143 0.8571429
#>  [237,] 0.2857143 0.8571429
#>  [238,] 0.1428571 0.8571429
#>  [239,] 0.1428571 0.8571429
#>  [240,] 0.1428571 0.7142857
#>  [241,] 0.2857143 0.7142857
#>  [242,] 0.1428571 0.8571429
#>  [243,] 0.2857143 0.7142857
#>  [244,] 0.1428571 0.8571429
#>  [245,] 0.1428571 0.8571429
#>  [246,] 0.2857143 0.5714286
#>  [247,] 0.1428571 0.8571429
#>  [248,] 0.1428571 0.8571429
#>  [249,] 0.1428571 0.8571429
#>  [250,] 0.5714286 0.8571429
#>  [251,] 0.2857143 0.8571429
#>  [252,] 0.2857143 0.8571429
#>  [253,] 0.1428571 0.8571429
#>  [254,] 0.2857143 0.8571429
#>  [255,] 0.2857143 0.8571429
#>  [256,] 0.2857143 0.8571429
#>  [257,] 0.1428571 0.7142857
#>  [258,] 0.2857143 0.7142857
#>  [259,] 0.1428571 0.8571429
#>  [260,] 0.1428571 0.7142857
#>  [261,] 0.1428571 0.8571429
#>  [262,] 0.1428571 0.8571429
#>  [263,] 0.1428571 0.8571429
#>  [264,] 0.1428571 0.7142857
#>  [265,] 0.4285714 0.8571429
#>  [266,] 0.1428571 0.5714286
#>  [267,] 0.2857143 0.7142857
#>  [268,] 0.2857143 0.7142857
#>  [269,] 0.1428571 0.7142857
#>  [270,] 0.1428571 0.8571429
#>  [271,] 0.4285714 0.8571429
#>  [272,] 0.2857143 0.7142857
#>  [273,] 0.1428571 0.8571429
#>  [274,] 0.2857143 0.8571429
#>  [275,] 0.1428571 0.8571429
#>  [276,] 0.5714286 0.8571429
#>  [277,] 0.2857143 0.8571429
#>  [278,] 0.2857143 0.8571429
#>  [279,] 0.1428571 0.8571429
#>  [280,] 0.1428571 0.8571429
#>  [281,] 0.1428571 0.8571429
#>  [282,] 0.2857143 0.7142857
#>  [283,] 0.2857143 0.8571429
#>  [284,] 0.1428571 0.8571429
#>  [285,] 0.1428571 0.7142857
#>  [286,] 0.1428571 0.7142857
#>  [287,] 0.1428571 0.8571429
#>  [288,] 0.1428571 0.8571429
#>  [289,] 0.1428571 0.8571429
#>  [290,] 0.1428571 0.7142857
#>  [291,] 0.1428571 0.8571429
#>  [292,] 0.1428571 0.8571429
#>  [293,] 0.2857143 0.7142857
#>  [294,] 0.1428571 0.7142857
#>  [295,] 0.1428571 0.5714286
#>  [296,] 0.1428571 0.8571429
#>  [297,] 0.4285714 0.8571429
#>  [298,] 0.1428571 0.7142857
#>  [299,] 0.1428571 0.8571429
#>  [300,] 0.1428571 0.8571429
#>  [301,] 0.2857143 0.8571429
#>  [302,] 0.1428571 0.7142857
#>  [303,] 0.1428571 0.8571429
#>  [304,] 0.2857143 0.8571429
#>  [305,] 0.1428571 0.8571429
#>  [306,] 0.1428571 0.7142857
#>  [307,] 0.2857143 0.7142857
#>  [308,] 0.2857143 0.8571429
#>  [309,] 0.1428571 0.7142857
#>  [310,] 0.1428571 0.8571429
#>  [311,] 0.1428571 0.8571429
#>  [312,] 0.2857143 0.5714286
#>  [313,] 0.2857143 0.8571429
#>  [314,] 0.1428571 0.8571429
#>  [315,] 0.4285714 0.8571429
#>  [316,] 0.1428571 0.7142857
#>  [317,] 0.1428571 0.8571429
#>  [318,] 0.1428571 0.5714286
#>  [319,] 0.2857143 0.8571429
#>  [320,] 0.4285714 0.7142857
#>  [321,] 0.1428571 0.8571429
#>  [322,] 0.1428571 0.8571429
#>  [323,] 0.2857143 0.7142857
#>  [324,] 0.1428571 0.8571429
#>  [325,] 0.1428571 0.8571429
#>  [326,] 0.2857143 0.8571429
#>  [327,] 0.2857143 0.8571429
#>  [328,] 0.1428571 0.8571429
#>  [329,] 0.1428571 0.7142857
#>  [330,] 0.1428571 0.8571429
#>  [331,] 0.2857143 0.8571429
#>  [332,] 0.2857143 0.8571429
#>  [333,] 0.2857143 0.8571429
#>  [334,] 0.1428571 0.5714286
#>  [335,] 0.2857143 0.8571429
#>  [336,] 0.2857143 0.7142857
#>  [337,] 0.1428571 0.8571429
#>  [338,] 0.1428571 0.7142857
#>  [339,] 0.1428571 0.5714286
#>  [340,] 0.1428571 0.7142857
#>  [341,] 0.1428571 0.8571429
#>  [342,] 0.1428571 0.7142857
#>  [343,] 0.2857143 0.8571429
#>  [344,] 0.1428571 0.8571429
#>  [345,] 0.1428571 0.8571429
#>  [346,] 0.1428571 0.8571429
#>  [347,] 0.1428571 0.8571429
#>  [348,] 0.1428571 0.7142857
#>  [349,] 0.1428571 0.7142857
#>  [350,] 0.1428571 0.8571429
#>  [351,] 0.1428571 0.8571429
#>  [352,] 0.2857143 0.5714286
#>  [353,] 0.2857143 0.8571429
#>  [354,] 0.2857143 0.8571429
#>  [355,] 0.1428571 0.7142857
#>  [356,] 0.1428571 0.8571429
#>  [357,] 0.1428571 0.8571429
#>  [358,] 0.2857143 0.7142857
#>  [359,] 0.1428571 0.8571429
#>  [360,] 0.1428571 0.8571429
#>  [361,] 0.1428571 0.8571429
#>  [362,] 0.4285714 0.7142857
#>  [363,] 0.2857143 0.7142857
#>  [364,] 0.1428571 0.8571429
#>  [365,] 0.1428571 0.7142857
#>  [366,] 0.2857143 0.8571429
#>  [367,] 0.1428571 0.8571429
#>  [368,] 0.1428571 0.8571429
#>  [369,] 0.2857143 0.8571429
#>  [370,] 0.1428571 0.7142857
#>  [371,] 0.1428571 0.5714286
#>  [372,] 0.1428571 0.7142857
#>  [373,] 0.1428571 0.8571429
#>  [374,] 0.1428571 0.8571429
#>  [375,] 0.1428571 0.8571429
#>  [376,] 0.2857143 0.5714286
#>  [377,] 0.1428571 0.7142857
#>  [378,] 0.2857143 0.5714286
#>  [379,] 0.2857143 0.8571429
#>  [380,] 0.1428571 0.7142857
#>  [381,] 0.1428571 0.5714286
#>  [382,] 0.1428571 0.8571429
#>  [383,] 0.1428571 0.8571429
#>  [384,] 0.1428571 0.8571429
#>  [385,] 0.2857143 0.8571429
#>  [386,] 0.2857143 0.8571429
#>  [387,] 0.2857143 0.7142857
#>  [388,] 0.1428571 0.8571429
#>  [389,] 0.2857143 0.8571429
#>  [390,] 0.2857143 0.8571429
#>  [391,] 0.1428571 0.7142857
#>  [392,] 0.1428571 0.7142857
#>  [393,] 0.1428571 0.8571429
#>  [394,] 0.1428571 0.8571429
#>  [395,] 0.1428571 0.8571429
#>  [396,] 0.2857143 0.8571429
#>  [397,] 0.1428571 0.8571429
#>  [398,] 0.1428571 0.8571429
#>  [399,] 0.1428571 0.5714286
#>  [400,] 0.1428571 0.7142857
#>  [401,] 0.2857143 0.7142857
#>  [402,] 0.1428571 0.4285714
#>  [403,] 0.1428571 0.7142857
#>  [404,] 0.2857143 0.4285714
#>  [405,] 0.2857143 0.8571429
#>  [406,] 0.1428571 0.8571429
#>  [407,] 0.1428571 0.8571429
#>  [408,] 0.1428571 0.5714286
#>  [409,] 0.4285714 0.8571429
#>  [410,] 0.1428571 0.8571429
#>  [411,] 0.4285714 0.8571429
#>  [412,] 0.1428571 0.8571429
#>  [413,] 0.1428571 0.8571429
#>  [414,] 0.2857143 0.7142857
#>  [415,] 0.1428571 0.8571429
#>  [416,] 0.2857143 0.8571429
#>  [417,] 0.1428571 0.8571429
#>  [418,] 0.1428571 0.8571429
#>  [419,] 0.1428571 0.7142857
#>  [420,] 0.1428571 0.8571429
#>  [421,] 0.2857143 0.8571429
#>  [422,] 0.1428571 0.7142857
#>  [423,] 0.1428571 0.8571429
#>  [424,] 0.1428571 0.8571429
#>  [425,] 0.1428571 0.7142857
#>  [426,] 0.2857143 0.7142857
#>  [427,] 0.2857143 0.7142857
#>  [428,] 0.1428571 0.8571429
#>  [429,] 0.2857143 0.8571429
#>  [430,] 0.1428571 0.7142857
#>  [431,] 0.4285714 0.7142857
#>  [432,] 0.4285714 0.7142857
#>  [433,] 0.1428571 0.5714286
#>  [434,] 0.2857143 0.7142857
#>  [435,] 0.1428571 0.7142857
#>  [436,] 0.1428571 0.7142857
#>  [437,] 0.2857143 0.7142857
#>  [438,] 0.1428571 0.8571429
#>  [439,] 0.1428571 0.8571429
#>  [440,] 0.1428571 0.7142857
#>  [441,] 0.1428571 0.8571429
#>  [442,] 0.2857143 0.8571429
#>  [443,] 0.2857143 0.8571429
#>  [444,] 0.2857143 0.7142857
#>  [445,] 0.2857143 0.7142857
#>  [446,] 0.2857143 0.7142857
#>  [447,] 0.2857143 0.8571429
#>  [448,] 0.4285714 0.8571429
#>  [449,] 0.1428571 0.8571429
#>  [450,] 0.2857143 0.8571429
#>  [451,] 0.1428571 0.8571429
#>  [452,] 0.1428571 0.8571429
#>  [453,] 0.1428571 0.8571429
#>  [454,] 0.1428571 0.7142857
#>  [455,] 0.2857143 0.8571429
#>  [456,] 0.1428571 0.8571429
#>  [457,] 0.1428571 0.8571429
#>  [458,] 0.2857143 0.8571429
#>  [459,] 0.1428571 0.7142857
#>  [460,] 0.1428571 0.8571429
#>  [461,] 0.1428571 0.8571429
#>  [462,] 0.1428571 0.8571429
#>  [463,] 0.1428571 0.8571429
#>  [464,] 0.2857143 0.8571429
#>  [465,] 0.1428571 0.8571429
#>  [466,] 0.1428571 0.8571429
#>  [467,] 0.2857143 0.8571429
#>  [468,] 0.2857143 0.8571429
#>  [469,] 0.1428571 0.7142857
#>  [470,] 0.4285714 0.7142857
#>  [471,] 0.2857143 0.8571429
#>  [472,] 0.2857143 0.8571429
#>  [473,] 0.1428571 0.8571429
#>  [474,] 0.1428571 0.8571429
#>  [475,] 0.1428571 0.7142857
#>  [476,] 0.1428571 0.8571429
#>  [477,] 0.4285714 0.5714286
#>  [478,] 0.1428571 0.7142857
#>  [479,] 0.1428571 0.8571429
#>  [480,] 0.2857143 0.8571429
#>  [481,] 0.1428571 0.8571429
#>  [482,] 0.4285714 0.8571429
#>  [483,] 0.2857143 0.8571429
#>  [484,] 0.1428571 0.7142857
#>  [485,] 0.1428571 0.8571429
#>  [486,] 0.2857143 0.5714286
#>  [487,] 0.2857143 0.7142857
#>  [488,] 0.2857143 0.5714286
#>  [489,] 0.2857143 0.8571429
#>  [490,] 0.1428571 0.8571429
#>  [491,] 0.1428571 0.8571429
#>  [492,] 0.1428571 0.8571429
#>  [493,] 0.1428571 0.8571429
#>  [494,] 0.2857143 0.7142857
#>  [495,] 0.2857143 0.8571429
#>  [496,] 0.1428571 0.8571429
#>  [497,] 0.2857143 0.8571429
#>  [498,] 0.2857143 0.7142857
#>  [499,] 0.1428571 0.7142857
#>  [500,] 0.1428571 0.8571429
#>  [501,] 0.2857143 0.8571429
#>  [502,] 0.1428571 0.8571429
#>  [503,] 0.1428571 0.7142857
#>  [504,] 0.2857143 0.8571429
#>  [505,] 0.1428571 0.8571429
#>  [506,] 0.1428571 0.7142857
#>  [507,] 0.1428571 0.7142857
#>  [508,] 0.2857143 0.7142857
#>  [509,] 0.2857143 0.8571429
#>  [510,] 0.1428571 0.8571429
#>  [511,] 0.2857143 0.8571429
#>  [512,] 0.2857143 0.7142857
#>  [513,] 0.2857143 0.5714286
#>  [514,] 0.1428571 0.7142857
#>  [515,] 0.2857143 0.5714286
#>  [516,] 0.2857143 0.8571429
#>  [517,] 0.2857143 0.7142857
#>  [518,] 0.2857143 0.8571429
#>  [519,] 0.2857143 0.7142857
#>  [520,] 0.2857143 0.8571429
#>  [521,] 0.2857143 0.8571429
#>  [522,] 0.1428571 0.5714286
#>  [523,] 0.1428571 0.8571429
#>  [524,] 0.1428571 0.5714286
#>  [525,] 0.1428571 0.8571429
#>  [526,] 0.2857143 0.8571429
#>  [527,] 0.1428571 0.8571429
#>  [528,] 0.2857143 0.7142857
#>  [529,] 0.2857143 0.8571429
#>  [530,] 0.1428571 0.8571429
#>  [531,] 0.2857143 0.8571429
#>  [532,] 0.1428571 0.7142857
#>  [533,] 0.1428571 0.7142857
#>  [534,] 0.1428571 0.7142857
#>  [535,] 0.2857143 0.8571429
#>  [536,] 0.5714286 0.8571429
#>  [537,] 0.1428571 0.7142857
#>  [538,] 0.1428571 0.8571429
#>  [539,] 0.2857143 0.8571429
#>  [540,] 0.2857143 0.8571429
#>  [541,] 0.1428571 0.8571429
#>  [542,] 0.1428571 0.8571429
#>  [543,] 0.2857143 0.8571429
#>  [544,] 0.1428571 0.7142857
#>  [545,] 0.1428571 0.7142857
#>  [546,] 0.1428571 0.8571429
#>  [547,] 0.2857143 0.7142857
#>  [548,] 0.1428571 0.7142857
#>  [549,] 0.4285714 0.8571429
#>  [550,] 0.1428571 0.5714286
#>  [551,] 0.2857143 0.8571429
#>  [552,] 0.1428571 0.7142857
#>  [553,] 0.2857143 0.7142857
#>  [554,] 0.1428571 0.8571429
#>  [555,] 0.1428571 0.8571429
#>  [556,] 0.2857143 0.7142857
#>  [557,] 0.1428571 0.8571429
#>  [558,] 0.1428571 0.8571429
#>  [559,] 0.1428571 0.8571429
#>  [560,] 0.1428571 0.8571429
#>  [561,] 0.2857143 0.8571429
#>  [562,] 0.2857143 0.8571429
#>  [563,] 0.2857143 0.8571429
#>  [564,] 0.2857143 0.5714286
#>  [565,] 0.4285714 0.8571429
#>  [566,] 0.2857143 0.7142857
#>  [567,] 0.1428571 0.7142857
#>  [568,] 0.1428571 0.7142857
#>  [569,] 0.1428571 0.8571429
#>  [570,] 0.2857143 0.5714286
#>  [571,] 0.1428571 0.8571429
#>  [572,] 0.1428571 0.8571429
#>  [573,] 0.4285714 0.8571429
#>  [574,] 0.1428571 0.8571429
#>  [575,] 0.2857143 0.5714286
#>  [576,] 0.1428571 0.7142857
#>  [577,] 0.2857143 0.8571429
#>  [578,] 0.1428571 0.8571429
#>  [579,] 0.1428571 0.8571429
#>  [580,] 0.2857143 0.8571429
#>  [581,] 0.2857143 0.7142857
#>  [582,] 0.1428571 0.8571429
#>  [583,] 0.2857143 0.7142857
#>  [584,] 0.1428571 0.5714286
#>  [585,] 0.1428571 0.8571429
#>  [586,] 0.5714286 0.8571429
#>  [587,] 0.4285714 0.8571429
#>  [588,] 0.2857143 0.8571429
#>  [589,] 0.2857143 0.5714286
#>  [590,] 0.2857143 0.7142857
#>  [591,] 0.1428571 0.7142857
#>  [592,] 0.1428571 0.8571429
#>  [593,] 0.4285714 0.8571429
#>  [594,] 0.1428571 0.8571429
#>  [595,] 0.1428571 0.8571429
#>  [596,] 0.1428571 0.8571429
#>  [597,] 0.2857143 0.5714286
#>  [598,] 0.2857143 0.5714286
#>  [599,] 0.1428571 0.7142857
#>  [600,] 0.2857143 0.8571429
#>  [601,] 0.1428571 0.8571429
#>  [602,] 0.1428571 0.8571429
#>  [603,] 0.1428571 0.8571429
#>  [604,] 0.4285714 0.8571429
#>  [605,] 0.1428571 0.7142857
#>  [606,] 0.1428571 0.8571429
#>  [607,] 0.1428571 0.8571429
#>  [608,] 0.2857143 0.8571429
#>  [609,] 0.2857143 0.8571429
#>  [610,] 0.2857143 0.8571429
#>  [611,] 0.1428571 0.8571429
#>  [612,] 0.1428571 0.7142857
#>  [613,] 0.1428571 0.7142857
#>  [614,] 0.1428571 0.5714286
#>  [615,] 0.2857143 0.8571429
#>  [616,] 0.1428571 0.8571429
#>  [617,] 0.2857143 0.7142857
#>  [618,] 0.1428571 0.8571429
#>  [619,] 0.4285714 0.8571429
#>  [620,] 0.2857143 0.7142857
#>  [621,] 0.1428571 0.5714286
#>  [622,] 0.1428571 0.7142857
#>  [623,] 0.1428571 0.7142857
#>  [624,] 0.1428571 0.8571429
#>  [625,] 0.1428571 0.7142857
#>  [626,] 0.1428571 0.8571429
#>  [627,] 0.1428571 0.8571429
#>  [628,] 0.4285714 0.7142857
#>  [629,] 0.2857143 0.7142857
#>  [630,] 0.1428571 0.8571429
#>  [631,] 0.2857143 0.8571429
#>  [632,] 0.2857143 0.8571429
#>  [633,] 0.1428571 0.8571429
#>  [634,] 0.1428571 0.7142857
#>  [635,] 0.2857143 0.5714286
#>  [636,] 0.1428571 0.8571429
#>  [637,] 0.2857143 0.8571429
#>  [638,] 0.2857143 0.5714286
#>  [639,] 0.2857143 0.8571429
#>  [640,] 0.2857143 0.7142857
#>  [641,] 0.2857143 0.8571429
#>  [642,] 0.1428571 0.7142857
#>  [643,] 0.4285714 0.7142857
#>  [644,] 0.4285714 0.8571429
#>  [645,] 0.1428571 0.7142857
#>  [646,] 0.1428571 0.7142857
#>  [647,] 0.1428571 0.8571429
#>  [648,] 0.4285714 0.4285714
#>  [649,] 0.4285714 0.7142857
#>  [650,] 0.1428571 0.8571429
#>  [651,] 0.2857143 0.8571429
#>  [652,] 0.1428571 0.7142857
#>  [653,] 0.1428571 0.8571429
#>  [654,] 0.1428571 0.8571429
#>  [655,] 0.1428571 0.7142857
#>  [656,] 0.1428571 0.4285714
#>  [657,] 0.4285714 0.8571429
#>  [658,] 0.2857143 0.5714286
#>  [659,] 0.1428571 0.8571429
#>  [660,] 0.1428571 0.8571429
#>  [661,] 0.1428571 0.5714286
#>  [662,] 0.4285714 0.5714286
#>  [663,] 0.2857143 0.7142857
#>  [664,] 0.2857143 0.8571429
#>  [665,] 0.2857143 0.8571429
#>  [666,] 0.2857143 0.8571429
#>  [667,] 0.2857143 0.8571429
#>  [668,] 0.2857143 0.8571429
#>  [669,] 0.4285714 0.8571429
#>  [670,] 0.2857143 0.8571429
#>  [671,] 0.1428571 0.8571429
#>  [672,] 0.4285714 0.7142857
#>  [673,] 0.1428571 0.7142857
#>  [674,] 0.1428571 0.7142857
#>  [675,] 0.1428571 0.8571429
#>  [676,] 0.2857143 0.8571429
#>  [677,] 0.1428571 0.8571429
#>  [678,] 0.2857143 0.7142857
#>  [679,] 0.2857143 0.8571429
#>  [680,] 0.1428571 0.7142857
#>  [681,] 0.1428571 0.8571429
#>  [682,] 0.1428571 0.8571429
#>  [683,] 0.1428571 0.8571429
#>  [684,] 0.1428571 0.7142857
#>  [685,] 0.4285714 0.8571429
#>  [686,] 0.1428571 0.8571429
#>  [687,] 0.2857143 0.8571429
#>  [688,] 0.2857143 0.8571429
#>  [689,] 0.1428571 0.8571429
#>  [690,] 0.4285714 0.7142857
#>  [691,] 0.2857143 0.8571429
#>  [692,] 0.2857143 0.8571429
#>  [693,] 0.1428571 0.8571429
#>  [694,] 0.2857143 0.7142857
#>  [695,] 0.2857143 0.5714286
#>  [696,] 0.2857143 0.5714286
#>  [697,] 0.2857143 0.8571429
#>  [698,] 0.2857143 0.8571429
#>  [699,] 0.1428571 0.7142857
#>  [700,] 0.1428571 0.7142857
#>  [701,] 0.1428571 0.8571429
#>  [702,] 0.2857143 0.8571429
#>  [703,] 0.1428571 0.8571429
#>  [704,] 0.1428571 0.7142857
#>  [705,] 0.1428571 0.7142857
#>  [706,] 0.2857143 0.8571429
#>  [707,] 0.1428571 0.7142857
#>  [708,] 0.1428571 0.8571429
#>  [709,] 0.1428571 0.8571429
#>  [710,] 0.1428571 0.8571429
#>  [711,] 0.2857143 0.8571429
#>  [712,] 0.1428571 0.7142857
#>  [713,] 0.2857143 0.8571429
#>  [714,] 0.1428571 0.8571429
#>  [715,] 0.4285714 0.7142857
#>  [716,] 0.1428571 0.7142857
#>  [717,] 0.2857143 0.7142857
#>  [718,] 0.2857143 0.7142857
#>  [719,] 0.1428571 0.7142857
#>  [720,] 0.1428571 0.8571429
#>  [721,] 0.1428571 0.8571429
#>  [722,] 0.1428571 0.7142857
#>  [723,] 0.4285714 0.8571429
#>  [724,] 0.1428571 0.5714286
#>  [725,] 0.2857143 0.8571429
#>  [726,] 0.4285714 0.7142857
#>  [727,] 0.2857143 0.8571429
#>  [728,] 0.1428571 0.7142857
#>  [729,] 0.2857143 0.7142857
#>  [730,] 0.1428571 0.8571429
#>  [731,] 0.2857143 0.8571429
#>  [732,] 0.1428571 0.8571429
#>  [733,] 0.1428571 0.7142857
#>  [734,] 0.2857143 0.5714286
#>  [735,] 0.1428571 0.7142857
#>  [736,] 0.1428571 0.7142857
#>  [737,] 0.2857143 0.7142857
#>  [738,] 0.2857143 0.8571429
#>  [739,] 0.2857143 0.7142857
#>  [740,] 0.4285714 0.7142857
#>  [741,] 0.4285714 0.8571429
#>  [742,] 0.2857143 0.5714286
#>  [743,] 0.2857143 0.8571429
#>  [744,] 0.1428571 0.7142857
#>  [745,] 0.1428571 0.7142857
#>  [746,] 0.1428571 0.7142857
#>  [747,] 0.1428571 0.5714286
#>  [748,] 0.1428571 0.8571429
#>  [749,] 0.2857143 0.5714286
#>  [750,] 0.1428571 0.8571429
#>  [751,] 0.1428571 0.7142857
#>  [752,] 0.2857143 0.7142857
#>  [753,] 0.2857143 0.8571429
#>  [754,] 0.2857143 0.5714286
#>  [755,] 0.1428571 0.8571429
#>  [756,] 0.1428571 0.8571429
#>  [757,] 0.1428571 0.8571429
#>  [758,] 0.2857143 0.8571429
#>  [759,] 0.1428571 0.8571429
#>  [760,] 0.1428571 0.8571429
#>  [761,] 0.1428571 0.8571429
#>  [762,] 0.4285714 0.8571429
#>  [763,] 0.1428571 0.7142857
#>  [764,] 0.1428571 0.8571429
#>  [765,] 0.4285714 0.8571429
#>  [766,] 0.1428571 0.8571429
#>  [767,] 0.1428571 0.8571429
#>  [768,] 0.1428571 0.7142857
#>  [769,] 0.2857143 0.8571429
#>  [770,] 0.4285714 0.8571429
#>  [771,] 0.1428571 0.8571429
#>  [772,] 0.1428571 0.8571429
#>  [773,] 0.1428571 0.8571429
#>  [774,] 0.2857143 0.7142857
#>  [775,] 0.1428571 0.8571429
#>  [776,] 0.1428571 0.8571429
#>  [777,] 0.1428571 0.8571429
#>  [778,] 0.1428571 0.8571429
#>  [779,] 0.2857143 0.8571429
#>  [780,] 0.1428571 0.8571429
#>  [781,] 0.1428571 0.7142857
#>  [782,] 0.1428571 0.8571429
#>  [783,] 0.1428571 0.8571429
#>  [784,] 0.1428571 0.7142857
#>  [785,] 0.2857143 0.8571429
#>  [786,] 0.1428571 0.8571429
#>  [787,] 0.1428571 0.5714286
#>  [788,] 0.1428571 0.8571429
#>  [789,] 0.1428571 0.7142857
#>  [790,] 0.2857143 0.8571429
#>  [791,] 0.1428571 0.8571429
#>  [792,] 0.2857143 0.7142857
#>  [793,] 0.2857143 0.8571429
#>  [794,] 0.1428571 0.7142857
#>  [795,] 0.2857143 0.8571429
#>  [796,] 0.2857143 0.7142857
#>  [797,] 0.4285714 0.5714286
#>  [798,] 0.1428571 0.8571429
#>  [799,] 0.2857143 0.8571429
#>  [800,] 0.1428571 0.7142857
#>  [801,] 0.1428571 0.8571429
#>  [802,] 0.1428571 0.8571429
#>  [803,] 0.1428571 0.8571429
#>  [804,] 0.1428571 0.8571429
#>  [805,] 0.2857143 0.7142857
#>  [806,] 0.2857143 0.8571429
#>  [807,] 0.1428571 0.8571429
#>  [808,] 0.1428571 0.7142857
#>  [809,] 0.1428571 0.7142857
#>  [810,] 0.1428571 0.7142857
#>  [811,] 0.2857143 0.8571429
#>  [812,] 0.1428571 0.8571429
#>  [813,] 0.1428571 0.7142857
#>  [814,] 0.2857143 0.7142857
#>  [815,] 0.2857143 0.8571429
#>  [816,] 0.2857143 0.8571429
#>  [817,] 0.1428571 0.7142857
#>  [818,] 0.1428571 0.8571429
#>  [819,] 0.2857143 0.7142857
#>  [820,] 0.2857143 0.7142857
#>  [821,] 0.2857143 0.8571429
#>  [822,] 0.2857143 0.8571429
#>  [823,] 0.1428571 0.7142857
#>  [824,] 0.1428571 0.8571429
#>  [825,] 0.1428571 0.8571429
#>  [826,] 0.2857143 0.8571429
#>  [827,] 0.2857143 0.7142857
#>  [828,] 0.1428571 0.7142857
#>  [829,] 0.1428571 0.7142857
#>  [830,] 0.4285714 0.8571429
#>  [831,] 0.1428571 0.7142857
#>  [832,] 0.2857143 0.7142857
#>  [833,] 0.1428571 0.8571429
#>  [834,] 0.1428571 0.8571429
#>  [835,] 0.1428571 0.7142857
#>  [836,] 0.1428571 0.7142857
#>  [837,] 0.2857143 0.7142857
#>  [838,] 0.2857143 0.8571429
#>  [839,] 0.4285714 0.8571429
#>  [840,] 0.1428571 0.8571429
#>  [841,] 0.4285714 0.7142857
#>  [842,] 0.2857143 0.8571429
#>  [843,] 0.1428571 0.8571429
#>  [844,] 0.1428571 0.5714286
#>  [845,] 0.1428571 0.8571429
#>  [846,] 0.1428571 0.8571429
#>  [847,] 0.2857143 0.7142857
#>  [848,] 0.1428571 0.7142857
#>  [849,] 0.1428571 0.8571429
#>  [850,] 0.1428571 0.8571429
#>  [851,] 0.1428571 0.8571429
#>  [852,] 0.1428571 0.8571429
#>  [853,] 0.1428571 0.7142857
#>  [854,] 0.2857143 0.8571429
#>  [855,] 0.2857143 0.8571429
#>  [856,] 0.1428571 0.8571429
#>  [857,] 0.2857143 0.7142857
#>  [858,] 0.1428571 0.8571429
#>  [859,] 0.1428571 0.8571429
#>  [860,] 0.2857143 0.8571429
#>  [861,] 0.2857143 0.8571429
#>  [862,] 0.2857143 0.8571429
#>  [863,] 0.4285714 0.8571429
#>  [864,] 0.4285714 0.8571429
#>  [865,] 0.2857143 0.8571429
#>  [866,] 0.5714286 0.8571429
#>  [867,] 0.2857143 0.7142857
#>  [868,] 0.1428571 0.7142857
#>  [869,] 0.1428571 0.7142857
#>  [870,] 0.2857143 0.8571429
#>  [871,] 0.2857143 0.7142857
#>  [872,] 0.1428571 0.8571429
#>  [873,] 0.2857143 0.8571429
#>  [874,] 0.1428571 0.7142857
#>  [875,] 0.1428571 0.7142857
#>  [876,] 0.1428571 0.8571429
#>  [877,] 0.2857143 0.8571429
#>  [878,] 0.2857143 0.8571429
#>  [879,] 0.1428571 0.8571429
#>  [880,] 0.1428571 0.8571429
#>  [881,] 0.2857143 0.8571429
#>  [882,] 0.1428571 0.8571429
#>  [883,] 0.1428571 0.8571429
#>  [884,] 0.1428571 0.8571429
#>  [885,] 0.2857143 0.7142857
#>  [886,] 0.1428571 0.4285714
#>  [887,] 0.1428571 0.5714286
#>  [888,] 0.4285714 0.7142857
#>  [889,] 0.2857143 0.8571429
#>  [890,] 0.1428571 0.7142857
#>  [891,] 0.2857143 0.7142857
#>  [892,] 0.1428571 0.8571429
#>  [893,] 0.1428571 0.5714286
#>  [894,] 0.2857143 0.5714286
#>  [895,] 0.1428571 0.7142857
#>  [896,] 0.2857143 0.7142857
#>  [897,] 0.2857143 0.7142857
#>  [898,] 0.1428571 0.8571429
#>  [899,] 0.1428571 0.8571429
#>  [900,] 0.2857143 0.8571429
#>  [901,] 0.4285714 0.8571429
#>  [902,] 0.1428571 0.8571429
#>  [903,] 0.2857143 0.8571429
#>  [904,] 0.1428571 0.7142857
#>  [905,] 0.2857143 0.7142857
#>  [906,] 0.1428571 0.8571429
#>  [907,] 0.1428571 0.8571429
#>  [908,] 0.4285714 0.8571429
#>  [909,] 0.2857143 0.7142857
#>  [910,] 0.1428571 0.8571429
#>  [911,] 0.1428571 0.8571429
#>  [912,] 0.4285714 0.8571429
#>  [913,] 0.2857143 0.8571429
#>  [914,] 0.1428571 0.8571429
#>  [915,] 0.2857143 0.5714286
#>  [916,] 0.1428571 0.8571429
#>  [917,] 0.1428571 0.8571429
#>  [918,] 0.1428571 0.7142857
#>  [919,] 0.1428571 0.8571429
#>  [920,] 0.2857143 0.5714286
#>  [921,] 0.1428571 0.8571429
#>  [922,] 0.1428571 0.8571429
#>  [923,] 0.2857143 0.8571429
#>  [924,] 0.1428571 0.7142857
#>  [925,] 0.2857143 0.8571429
#>  [926,] 0.2857143 0.7142857
#>  [927,] 0.2857143 0.8571429
#>  [928,] 0.1428571 0.8571429
#>  [929,] 0.1428571 0.7142857
#>  [930,] 0.1428571 0.7142857
#>  [931,] 0.1428571 0.8571429
#>  [932,] 0.1428571 0.8571429
#>  [933,] 0.4285714 0.8571429
#>  [934,] 0.2857143 0.8571429
#>  [935,] 0.1428571 0.8571429
#>  [936,] 0.1428571 0.8571429
#>  [937,] 0.4285714 0.8571429
#>  [938,] 0.2857143 0.7142857
#>  [939,] 0.1428571 0.8571429
#>  [940,] 0.1428571 0.8571429
#>  [941,] 0.1428571 0.8571429
#>  [942,] 0.2857143 0.8571429
#>  [943,] 0.2857143 0.8571429
#>  [944,] 0.2857143 0.8571429
#>  [945,] 0.2857143 0.8571429
#>  [946,] 0.1428571 0.7142857
#>  [947,] 0.2857143 0.8571429
#>  [948,] 0.1428571 0.7142857
#>  [949,] 0.2857143 0.8571429
#>  [950,] 0.1428571 0.7142857
#>  [951,] 0.1428571 0.8571429
#>  [952,] 0.1428571 0.7142857
#>  [953,] 0.1428571 0.8571429
#>  [954,] 0.1428571 0.7142857
#>  [955,] 0.1428571 0.8571429
#>  [956,] 0.1428571 0.8571429
#>  [957,] 0.2857143 0.5714286
#>  [958,] 0.1428571 0.5714286
#>  [959,] 0.4285714 0.8571429
#>  [960,] 0.1428571 0.8571429
#>  [961,] 0.2857143 0.8571429
#>  [962,] 0.2857143 0.8571429
#>  [963,] 0.1428571 0.7142857
#>  [964,] 0.2857143 0.8571429
#>  [965,] 0.1428571 0.8571429
#>  [966,] 0.1428571 0.8571429
#>  [967,] 0.1428571 0.8571429
#>  [968,] 0.1428571 0.7142857
#>  [969,] 0.1428571 0.8571429
#>  [970,] 0.4285714 0.8571429
#>  [971,] 0.1428571 0.8571429
#>  [972,] 0.2857143 0.8571429
#>  [973,] 0.4285714 0.8571429
#>  [974,] 0.1428571 0.5714286
#>  [975,] 0.4285714 0.7142857
#>  [976,] 0.1428571 0.7142857
#>  [977,] 0.2857143 0.8571429
#>  [978,] 0.1428571 0.8571429
#>  [979,] 0.1428571 0.8571429
#>  [980,] 0.2857143 0.8571429
#>  [981,] 0.2857143 0.8571429
#>  [982,] 0.1428571 0.8571429
#>  [983,] 0.1428571 0.5714286
#>  [984,] 0.1428571 0.7142857
#>  [985,] 0.1428571 0.5714286
#>  [986,] 0.1428571 0.8571429
#>  [987,] 0.1428571 0.8571429
#>  [988,] 0.2857143 0.8571429
#>  [989,] 0.1428571 0.8571429
#>  [990,] 0.2857143 0.8571429
#>  [991,] 0.1428571 0.8571429
#>  [992,] 0.1428571 0.8571429
#>  [993,] 0.1428571 0.8571429
#>  [994,] 0.1428571 0.7142857
#>  [995,] 0.2857143 0.8571429
#>  [996,] 0.2857143 0.8571429
#>  [997,] 0.2857143 0.7142857
#>  [998,] 0.1428571 0.8571429
#>  [999,] 0.4285714 0.8571429
#> [1000,] 0.4285714 0.7142857